Move SDMX execute into base class
